Add a flag indicating restart files, make WRFDA quit if given a restart file (#269)
TYPE: enhancement

KEYWORDS: Restart, output_wrf, WRFDA, FLAG_RESTART

SOURCE: internal

DESCRIPTION OF CHANGES: Restart files can not be used by WRFDA because (among other reasons) the variable names are not consistent between restart and wrfinput/wrfout files. However, these files will still be happily read and used by WRFDA, with garbage data being passed around (for sometimes a very long time, depending on the user's settings) until WRFDA eventually produces an output full of NaNs (or other garbage data), or crashes upon trying to produce the output.

There is currently no easy, non-hacky way to determine whether the file given is a restart file or not. Since this could be useful for other reasons beyond this WRFDA application, this pull request adds a flag integer "FLAG_RESTART" with a value of 1 to restart files when they are written. In WRFDA, this variable is searched for in the input file, and if it exists and is set to 1, WRFDA exits gracefully with an error message instructing the user that restart files can not be used.

This test will only affect files created with newer versions of WRF, running newer versions of WRFDA. For files created by older versions, WRFDA will still fail in the same way as before.

LIST OF MODIFIED FILES:
M share/output_wrf.F
M var/da/da_main/da_med_initialdata_input.inc

TESTS CONDUCTED: WRFDA regression test passes for GNU and Intel. WTF done and successful. Created a test WRF and WRFDA case for all 4 combinations of old/new WRF and WRFDA. For cases where WRF and or WRFDA did not have these new commits, WRFDA failed in the old manner: ugly (for this particular case, a wrfvar_output was created, with a bunch of fields full of NaNs). For the case where both WRF and WRFDA were updated with the changes in this pull request, WRFDA quits almost immediately with a helpful message:

fg appears to be a WRF restart file
WRFDA can only work with wrfinput or wrfout format files