Fix MYJ for use with LES (#522)
TYPE: bug fix

KEYWORDS: MYJ SFC with LES

SOURCE: internal (problem from user)

DESCRIPTION OF CHANGES:
LES requires array USTM which is u* for momentum. In SFCLAY this is different from UST because it excludes the convective velocity, and this is what is needed for LES. After MYJSFC we simply copy UST into USTM because it is used for momentum with the convective velocity effect. USTM not available in NMM so ifdef'd out.

LIST OF MODIFIED FILES:
M phys/module_surface_driver.F

TESTS CONDUCTED:
WTF with gnu passed all.
User confirmed it works with LES.