Add sf_ocean_physics to options needed to be turned off during DFI (#506)
TYPE: enhancement

KEYWORDS: dfi, sf_ocean_physics

SOURCE: internal

DESCRIPTION OF CHANGES:
This PR adds sf_ocean_physics to a bunch of options that are turned off during backward integration and on when integration is going forward.

LIST OF MODIFIED FILES:
M share/dfi.F

TESTS CONDUCTED:
Test conducted and it shows the code works as expected. If forecast continues after DFI, the option is turned back on during the forecast period. When one does this, it's better to set frame_per_outfiles to 1 so that history files for backward and forward integration are separated, or do not write any history files during DFI.