Part 1: prepare new location where most top-level READMEs will exist (#575)
### TYPE: text only ###

### KEYWORDS: READMEs ###

### SOURCE: internal ###

### DESCRIPTION OF CHANGES: ###
#### Problem: ####
The WRF top-level directory is getting cluttered, and one of the problems is the proliferation of README files.
#### Solution: #### 
We certainly do not want to get rid of the README files. Also, trying to locate the files in appropriate yet individually separate directories does not seem feasible. For example, README.DA could go in the var directory, but where would README.rsl_output go? The idea is to have a new directory specifically for README files. All of the READMEs in the top-level dir would go into this new holding area, except for a single README file. This single file would contain the usual release information, PUBLIC DOMAIN notice, and point to where the remainder of the README files are located. The first part of this two-part commit, this PR, is to just create the new holding directory (info) and modify the (soon to be single remaining) top-level README. The second part of the commit will move the README files to the info directory.

1. Make a new top-level directory, info.
2. In the info directory, add the file INSTRUCTIONS. When developers add a new README file, they also need to update the sole remaining top-level README file.

The naming convention for the new directory was considered.
1. We do not want a top-level file called README, and some directory called README_DIR (or something similar, with a suffix identifying the holding directory of the README files).
2. We cannot have a README file (uppercase) and also a readme directory (lowercase) because of Mac file systems.
3. Calling the directory "doc" or "DOCS" seemed an over-statement for what was really contained in the directory.

### LIST OF MODIFIED FILES: ### 
A	doc/INSTRUCTIONS
M	README

### TESTS CONDUCTED: ###
 - [x] No testing required