Add CODEOWNERS file to .github directory  (#658)
TYPE: new feature

KEYWORDS: CODEOWNERS

SOURCE: internal

DESCRIPTION OF CHANGES:
This is the start of a list for shared and individual responsibility
for files and directories. When a PR is issued that modifies one of
the known files or files within a known directory (and required
reviews are enabled on that branch), then an automatic request for
review is issued to the code owners.

CODEOWNERS file is linked to the GitHub via the following teams:
   * physics
   * shared_physics
   * da
   * chem
   * hwrf_dev
   * arw_dev
   * infrastructure
   * global_owner
   * fire

Replaces PR #437 "Add CODEOWNERS file to .github directory". The fork and branch were deleted, so amending was not easy. 

More info:
```
https://help.github.com/articles/about-codeowners/
```

LIST OF MODIFIED FILES:
M	     .github/CODEOWNERS

TESTS CONDUCTED:
 - [x] Successfully forced a "code owner" to review before merge was permitted.