Bug fix for diff_6th_opt restart problem (#736)
TYPE: bug fix

KEYWORDS: diff_6th_opt, restart

SOURCE: internal, reported by Jorge Navarro Instituto de Geociencias (UCM-CSIC) of Spain

DESCRIPTION OF CHANGES:
When diff_6th_opt is turned on, restart fails in producing bit-for-bit results in real-data (reported) as well as idealized cases. This is due to the start and end indices too close to the domain boundaries when doing 6th order horizontal diffusions. This is a bug since this option was put in in V3.1. Because of the loop indices are changed, the results from any run that uses this option will also change.

LIST OF MODIFIED FILES: list of changed files:
M dyn_em/module_big_step_utilities_em.F

TESTS CONDUCTED:
No regression test yet. But tested with real-data case, quarter_ss case with periodic and open boundary conditions, and b_wave case with symmetric boundary condition in y, the restart gives bit-for-bit results.

RELEASE NOTE:
Fixed a non bit-for-bit restart problem with diff_6th_opt option. The error was due to the start and end indices too close to the domain boundaries when doing 6th order horizontal diffusions. Because the loop indices are changed, the results from any run that uses this option will also change. This was a bug since this option was added to the model in V3.1. Thanks for Jorge Navarro for the initial report that let to the discovering of the bug.