bug fix in GOCART gravitational settling (#714)
TYPE: bug fix

KEYWORDS: GOCART, WRF-Chem, dust and sea salt gravitational settling

SOURCE: Alexander Ukhov, KAUST

DESCRIPTION OF CHANGES:
Problem: It was found that the dust (and sea salt) mass balance is violated in routine _settling_ (module module_gocart_settling.F). In particular, the total mass of the dust (and sea salt) is increasing, when the dust (and sea salt) settles from the upper to the lower layers. In the code the dust (and sea salt) mixing ratios are transfered between the layers, which is not correct, since the air density is changing.

Solution: the discretization scheme was replaced to the one, which accouts for air density change and allows to conserve the dust (and sea salt) total mass.

LIST OF MODIFIED FILES: 
chem/module_gocart_settling.F

TESTS CONDUCTED: To validate our approach we developed a simple case: in the center of the domain (Middle East, 4500x4500 km2) we allocated a small region (100x100 km2), which emits dust for 10 hours. The simulation stops before the dust reachs the boundaries of the domain. No inflow of the dust from the boundaries as well.

Thereby, the dust mass balance is the following:
EMITTED DUST = (dust in the atmosphere) + (gravitationally settled dust) + (dry deposited dust)


 The obtained results demonstrate the dust mass balance before and after the correction.
![total](https://user-images.githubusercontent.com/5716976/49227727-d83c3c80-f3fa-11e8-980b-f8cdb63a4f99.png)

I have attached derivation of the scheme.
![derivation](https://user-images.githubusercontent.com/5716976/49231381-df674880-f402-11e8-898c-4066c0d48b91.jpg)