Jensen microphysics small qi bug fix for vapor growth check (#820)
TYPE: bug fix

KEYWORDS: jensen, mp_physics, qi, floating_point

SOURCE: internal

DESCRIPTION OF CHANGES: In the vapor growth check in Jensen microphysics, qi had the ability to be negative, which means that in the equation for ani(cc), there was potential to raise a negative number to an exponent of a non-logical fractional value, which caused a floating point error. This only stopped the model from running when compiled with a check for floating point traps. Inserting an if statement to prevent this from happening corrects the problem.

LIST OF MODIFIED FILES:
M phys/module_mp_jensen_ishmael.F

TESTS CONDUCTED: Verified that code builds and runs to completion, even when compiled with floating point trap checks.