bug fix for mynn PBL restarts - not giving bit-for-bit results (#970)
TYPE: bug fix

KEYWORDS: bug, MYNN, PBL, cldfra_bl, icloud_bl, Registry

SOURCE: internal

DESCRIPTION OF CHANGES: When using the MYNN PBL scheme, restarts did not give bit-for-bit
results. When using this scheme, the default setting for icloud_bl is 1 (on), but the variable used
(cldfra_bl) was not put in the restart stream. Simply adding this variable to the restart stream in
Registry.EM_COMMON corrects the problem.

LIST OF MODIFIED FILES:
M Registry/Registry.EM_COMMON

TESTS CONDUCTED:
Confirmed that the code builds, and the change corrects the non-bit-for-bit restart problem.

RELEASE NOTE: When using the MYNN PBL scheme, with icloud_bl = 1 (which is default), restarts did not give bit-for-bit results when compared to a non-restart run. This has been a problem since the option was introduced in V3.8, but is now corrected.