bug fix: NoahMP STEPWTD==0, divide by zero (#1129)
TYPE: bug fix

KEYWORDS: NoahMP, opt_run option

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
When running the NoahMP scheme with the ground water option (opt_run=5), the STEPWTD
variable is initialized to a non-zero value. However when opt_run!=5, then the STEPWTD
variable remains an uninitialized zero value. In the WRF source code, there are a couple of
places where mod(x,STEPWTD) occurs. This causes an immediate "divide by zero" error in
the first call to the surface driver.

Solution:
The IF test logic is re-arranged so that only when opt_run=5 (which means that STEPWTD
is non-zero) is there a second IF test (using mod(x,STEPWTD)) to compute whether
additional processing takes place.

This modification was made separately to the develop branch after the branch release-v4.1.4
was released. See #1101. The purpose of this modification
is to allow users access to the latest v4.1.* branch that works with NoahMP.

As this will be a single issue release, the README and inc/version_decl are also modified
in preparation for the release of WRF release-v4.1.5 with this PR.

LIST OF MODIFIED FILES:
M dyn_em/module_first_rk_step_part1.F
M phys/module_surface_driver.F
M README
M inc/version_decl

TESTS CONDUCTED:

Jenkins test PASS (positively tests NoahMP)
Regression test in NCAR-MMM classroom machine PASS.