Irrigation clean-up: syntax error, derived nml packaging, protect against div by 0 (#1149)
TYPE: bug fix

KEYWORDS: irrigation, drip, statement function

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem 1:
In module_irrigation.F, in subroutine sprinkler_irrigation is this line:
```
constants_irrigation=irr_freq*irr_daily_amount/(irr_num_hours*3600*rho(a,kms,b)*dz8w(a,kms,b)*100)
```
It is clear from the rest of the routine, that this is to be treated as a statement function.
1. The line is after all of the declarations and before the first executable statement.
2. The values of a, b, and kms (among others) are not yet defined.
3. The following line _tries_ to use it as a statement function.
```
qr_curr(a,kms,b)=qr_curr(a,kms,b)+irrigation(a,b)*constants_irrigation*dt
```
However, the GNU compiler is segfaulting on the definition of the statement function,
because the variables "a" and "b" (again, among others) are not yet assigned and are likely
garbage values. The GNU compiler is segfaulting because it does not recognize this line
as a proper statement function.

Statement functions were declared obsolescent in f95, but we have them in other routines.

Solution 1:
Abandon the statement function completely. The code is actually written to not need a statement
function.

Problem 2:
The user defined integer `irr_num_hours` can be zero, resulting in a divide by zero.

Solution 2:
In check_a_mundo, verify that the number of hours to irrigate is >= 0. This handles all three irrigation
options in a single location.

Problem 3:
The irrigation schemes are designed to mostly be activated on the inner-most domain. As with many
physics schemes, ON vs OFF causes troubles for nested domains when packaging of variables is
involved.

Solution 3:
A derived namelist option is utilized. If any domain has an irrigation option activated, then all domains 
have the packaged fields allocated.

LIST OF MODIFIED FILES:
modified:   module_irrigation.F
modified:   Registry/Registry.EM_COMMON
modified:   share/module_check_a_mundo.F 

TESTS CONDUCTED:
1. Jenkins is all PASS. 
2. Results are identical to previous, from developer.
3. With no mod, with the sprinkler irrigation option activated:
```
 &physics
 sf_surf_irr_scheme                  = 3,     3,     3,
```
We get the WRF output as:
```
WRF NUMBER OF TILES =   1

Program received signal SIGSEGV: Segmentation fault - invalid memory reference.

Backtrace for this error:
	at /wrf/WRF/phys/module_irrigation.f90:140
```
4. With the mods, the code runs to completion. 
5. With Registry mods for derived nml for irrigation allocation, following is also OK for 2 domains:
```
 &physics
 sf_surf_irr_scheme                  = 0,     3,     3,
```