Bug fixes in deng shallow cumulus parameterization (#1277)
TYPE: bug fix

KEYWORDS: Deng shcu, resolved hydrometeors, non-active updrafts, precipitation units, variable initialization, limit NT

SOURCE: Pedro A. Jimenez and Tim Juliano (NCAR/RAL), Eric Grimit and Mattew Wiley (Vaisala), and Xin-Zhong Liang
(University of Maryland)

DESCRIPTION OF CHANGES:
Problem:

The resolved hydrometeors are added to the unresolved hydrometeors inside Deng's parameterization. The radiation
driver adds the resolved hydrometeors to the unresolved ones causing to double counting the effects of the resolved
hydrometeors.
The unresolved clouds are not properly removed when the updraft is not active
Incorrect precipitation units
Incorrect initialization of variable radiusc. The variable is initialized inside an #if construct for the 1D model
If dt < 18 s NT is larger than 100 causing the code to look for values outside of array bounds for ainckfsa variable
Solution:

Do not add the resolved hydrometeors to the unresolved ones inside Deng's parameterization
Set cloud fraction and the mixing ratio of the unresolved hydrometeors to 0 if the updraft is not active.
Remove the conversion factor that was leading to the incorrect units.
Initialize radiusc outside of the #if construct so radiusc is always initialized.
Limit the maximum value of DT to 100
The following figure illustrates the impact of corrections 1 and 2. The time series of the mixing ratio now properly goes
to 0 if the updraft is not active.
image

LIST OF MODIFIED FILES:
M phys/module_shcu_deng.F

TESTS CONDUCTED:

We have run simulations with and without the fixes. The figure above illustrates that the clouds are now adequately
removed if there are not active updrafts.
Jenkins tests are all passing
RELEASE NOTE: Several fixes introduced to the Deng's shallow cumulus parameterization: 1) Do not add the resolved hydrometeors to the unresolved ones to avoid double counting, 2) remove clouds if the updraft is not active, 3) correct conversion factor for precipitation, 4) correct initialization of radiusc variable, and 5) limit variable NT to avoid array out of bounds.